tests: Extract manifest helpers from utils
