+++ tests/utils.go
@@ -38,7 +38,6 @@ import (
 	"net/http"
 	"os"
 	"os/exec"
-	"path"
 	"path/filepath"
 	"reflect"
 	"sort"
@@ -62,14 +61,12 @@ import (
 	"k8s.io/apimachinery/pkg/api/meta"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/rand"
-	"k8s.io/apimachinery/pkg/util/yaml"
 	"k8s.io/apimachinery/pkg/watch"
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
@@ -719,62 +716,6 @@ func CreateHostPathPvWithSizeAndStorageClass(osName, hostPath, size, sc string) 
 	return libnode.SchedulableNode
 }
 
-func GetListOfManifests(pathToManifestsDir string) []string {
-	var manifests []string
-	matchFileName := func(pattern, filename string) bool {
-		match, err := filepath.Match(pattern, filename)
-		if err != nil {
-			panic(err)
-		}
-		return match
-	}
-	err := filepath.Walk(pathToManifestsDir, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			fmt.Printf("ERROR: Can not access a path %q: %v\n", path, err)
-			return err
-		}
-		if !info.IsDir() && matchFileName("*.yaml", info.Name()) {
-			manifests = append(manifests, path)
-		}
-		return nil
-	})
-	if err != nil {
-		fmt.Printf("ERROR: Walking the path %q: %v\n", pathToManifestsDir, err)
-		panic(err)
-	}
-	return manifests
-}
-
-func ReadManifestYamlFile(pathToManifest string) []unstructured.Unstructured {
-	var objects []unstructured.Unstructured
-	stream, err := os.Open(pathToManifest)
-	util2.PanicOnError(err)
-
-	decoder := yaml.NewYAMLOrJSONDecoder(stream, 1024)
-	for {
-		obj := map[string]interface{}{}
-		err := decoder.Decode(&obj)
-		if err == io.EOF {
-			break
-		} else if err != nil {
-			panic(err)
-		}
-		if len(obj) == 0 {
-			continue
-		}
-		objects = append(objects, unstructured.Unstructured{Object: obj})
-	}
-	return objects
-}
-
-func isNamespaceScoped(kind schema.GroupVersionKind) bool {
-	switch kind.Kind {
-	case "ClusterRole", "ClusterRoleBinding":
-		return false
-	}
-	return true
-}
-
 func ServiceMonitorEnabled() bool {
 	virtClient, err := kubecli.GetKubevirtClient()
 	util2.PanicOnError(err)
@@ -801,56 +742,6 @@ func PrometheusRuleEnabled() bool {
 	}
 
 	return prometheusRuleEnabled
-}
-
-func composeResourceURI(object unstructured.Unstructured) string {
-	uri := "/api"
-	if object.GetAPIVersion() != "v1" {
-		uri += "s"
-	}
-	uri = path.Join(uri, object.GetAPIVersion())
-	if object.GetNamespace() != "" && isNamespaceScoped(object.GroupVersionKind()) {
-		uri = path.Join(uri, "namespaces", object.GetNamespace())
-	}
-	uri = path.Join(uri, strings.ToLower(object.GetKind()))
-	if !strings.HasSuffix(object.GetKind(), "s") {
-		uri += "s"
-	}
-	return uri
-}
-
-func ApplyRawManifest(object unstructured.Unstructured) error {
-	virtCli, err := kubecli.GetKubevirtClient()
-	util2.PanicOnError(err)
-
-	uri := composeResourceURI(object)
-	jsonbody, err := object.MarshalJSON()
-	util2.PanicOnError(err)
-	b, err := virtCli.CoreV1().RESTClient().Post().RequestURI(uri).Body(jsonbody).DoRaw(context.Background())
-	if err != nil {
-		fmt.Printf(fmt.Sprintf("ERROR: Can not apply %s\n", object))
-		panic(err)
-	}
-	status := unstructured.Unstructured{}
-	return json.Unmarshal(b, &status)
-}
-
-func DeleteRawManifest(object unstructured.Unstructured) error {
-	virtCli, err := kubecli.GetKubevirtClient()
-	util2.PanicOnError(err)
-
-	uri := composeResourceURI(object)
-	uri = path.Join(uri, object.GetName())
-
-	policy := metav1.DeletePropagationBackground
-	options := &metav1.DeleteOptions{PropagationPolicy: &policy}
-
-	result := virtCli.CoreV1().RESTClient().Delete().RequestURI(uri).Body(options).Do(context.Background())
-	if result.Error() != nil && !errors.IsNotFound(result.Error()) {
-		fmt.Printf(fmt.Sprintf("ERROR: Can not delete %s err: %#v %s\n", object.GetName(), result.Error(), object))
-		panic(err)
-	}
-	return nil
 }
 
 func DeleteConfigMap(name string) {
